refactor: ``GetCheckInResponse`` DTO 생성 및 적용
``GetCheckInResponse`` DTO를 정의하고 이를 적용하였습니다

Resolves: #1
