+++ src/main/kotlin/rainbowfriends/daramserverv2/domain/checkin/service/CheckedInMemberService.kt
@@ -1,7 +1,7 @@
 package rainbowfriends.daramserverv2.domain.checkin.service
 
-import rainbowfriends.daramserverv2.global.checkin.entity.CheckIn
+import rainbowfriends.daramserverv2.domain.checkin.dto.response.GetCheckInResponse
 
 interface CheckedInMemberService {
-    fun getCheckedInMember(): List<CheckIn>
+    fun getCheckedInMember(): List<GetCheckInResponse>
 }
