+++ src/main/kotlin/rainbowfriends/daramserverv2/domain/checkin/service/MissedCheckInMemberService.kt
@@ -1,7 +1,7 @@
 package rainbowfriends.daramserverv2.domain.checkin.service
 
-import rainbowfriends.daramserverv2.global.checkin.entity.CheckIn
+import rainbowfriends.daramserverv2.domain.checkin.dto.response.GetCheckInResponse
 
 interface MissedCheckInMemberService {
-    fun getMissedCheckInMember(): List<CheckIn>
+    fun getMissedCheckInMember(): List<GetCheckInResponse>
 }
